requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
dataclasses-json==0.6.5
distlib==0.3.8
distro==1.9.0
dnspython==1.16.0
dulwich==0.21.7
fastapi==0.110.3
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.1
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.48
langchain-openai==0.1.4
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.52
marshmallow==3.21.1
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.25.0
orjson==3.10.2
packaging==23.2
pexpect==4.9.0
pinecone-client==3.2.2
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==3.11.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.4.28
replicate==0.26.0
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tenacity==8.2.3
tiktoken==0.6.0
tomlkit==0.12.5
tqdm==4.66.2
trove-classifiers==2024.4.10
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.2
xattr==1.1.0
yarl==1.9.4
zipp==3.18.1
